+++ server/flags/flags.go
@@ -53,7 +53,8 @@ const (
 	// JSONRPCEnableMetrics enables EVM RPC metrics server.
 	// Set to `metrics` which is hardcoded flag from go-ethereum.
 	// https://github.com/ethereum/go-ethereum/blob/master/metrics/metrics.go#L35-L55
-	JSONRPCEnableMetrics = "metrics"
+	JSONRPCEnableMetrics   = "metrics"
+	JSONRPCReturnDataLimit = "json-rpc.return-data-limit"
 )
 
 // EVM flags